requirements.txt
requests>=2.25.1
pandas>=1.3.0
tabulate>=0.8.9
flask>=2.0.1
flask_session>=0.4.0
python-dotenv>=0.19.0